Initial Commit
This is the inital framework for the vehicle simulator.
The switch changes the mode that the device is operating
Optional Modes: doRun, doExtend, doContract, doDemo
doRun sends a unsigned char via CAN to the vehicle interface, based off
the reading of the display potentiometer.
doExtend set the actuator to max
doContract sets the actuator to min
doDemo toggles between doExtend and doContract every 5 seconds